Initial 0.3.0: checkpointing, more input formats, HTML report

+++ examples/remove_background/generate_tiny_10x_pbmc.py
@@ -2,17 +2,13 @@
 
 import urllib.request
 import sys
-import tarfile
 import os
 import numpy as np
-from scipy.io import mmread, mmwrite
-import pandas as pd
-import operator
-import shutil
+from cellbender.remove_background.downstream import load_anndata_from_input
 
 dataset_name = "pbmc4k (CellRanger 2.1.0, v2 Chemistry)"
-dataset_url = "http://cf.10xgenomics.com/samples/cell-exp/2.1.0/pbmc4k/pbmc4k_raw_gene_bc_matrices.tar.gz"
-dataset_local_filename = "pbmc4k_raw_gene_bc_matrices.tar.gz"
+dataset_url = "https://cf.10xgenomics.com/samples/cell-exp/2.1.0/pbmc4k/pbmc4k_raw_gene_bc_matrices_h5.h5"
+dataset_local_filename = "pbmc4k_raw_gene_bc_matrices_h5.h5"
 expected_cells = 4000
 num_cell_barcodes_to_keep = 500
 num_empty_barcodes_to_keep = 50_000
@@ -20,48 +16,43 @@ num_genes_to_keep = 100
 random_seed = 1984
 rng = np.random.RandomState(random_seed)
 
-print(f"Downloading {dataset_name}...")
-try:
-    urllib.request.urlretrieve(dataset_url, dataset_local_filename)
-except IOError:
-    print(f"Could not retrieve {dataset_name} -- cannot continue!")
-    sys.exit()
+if not os.path.exists(dataset_local_filename):
 
-print(f"Extracting {dataset_name}...")
-tar = tarfile.open(dataset_local_filename, "r:gz")
-tar.extractall()
-tar.close()
+    print(f"Downloading {dataset_name}...")
+    try:
+        urllib.request.urlretrieve(dataset_url, dataset_local_filename)
 
-extracted_dataset_local_path = os.path.join(os.curdir, "raw_gene_bc_matrices")
-matrix_mtx_path = os.path.join(extracted_dataset_local_path, "GRCh38", "matrix.mtx")
-genes_tsv_path = os.path.join(extracted_dataset_local_path, "GRCh38", "genes.tsv")
-barcodes_tsv_path = os.path.join(extracted_dataset_local_path, "GRCh38", "barcodes.tsv")
+    except urllib.error.HTTPError:
+        print(f"10x Genomics website is giving a 403 Forbidden error when an automatic "
+              f"download was attempted. Please visit "
+              f"https://www.10xgenomics.com/resources/datasets/4-k-pbm-cs-from-a-healthy-donor-2-standard-2-1-0 "
+              f"in a browser and manually download 'Gene / cell matrix HDF5 (raw)' into "
+              f"this folder and re-run this script.")
+        sys.exit()
 
-print("Loading the gene expression matrix...")
-matrix_mtx = mmread(matrix_mtx_path)
-num_raw_genes = matrix_mtx.shape[0]
-num_raw_barcodes = matrix_mtx.shape[1]
+    except IOError:
+        print(f"Could not retrieve {dataset_name} -- cannot continue!")
+        sys.exit()
 
-print("Loading genes and barcodes...")
-genes_df = pd.read_csv(genes_tsv_path, delimiter="\t", header=None)
-barcodes_df = pd.read_csv(barcodes_tsv_path, delimiter="\t", header=None)
+print("Loading the dataset...")
+adata = load_anndata_from_input(dataset_local_filename)
+num_raw_genes = adata.shape[1]
+print(f"Raw dataset size {adata.shape}")
 
 print(f"Trimming {dataset_name}...")
-# (naive) indices of expected cells
-umi_per_barcode = np.asarray(np.sum(matrix_mtx, 0)).flatten()
-total_gene_expression = np.asarray(np.sum(matrix_mtx, 1)).flatten()
-umi_sorted_barcode_indices = list(
-    map(operator.itemgetter(0),
-        sorted(enumerate(umi_per_barcode), key=operator.itemgetter(1), reverse=True)))
-cell_indices = umi_sorted_barcode_indices[:expected_cells]
-empty_indices = umi_sorted_barcode_indices[expected_cells:]
-
-# (naive) filter counts to non-empty droplets
-cell_counts_csr = matrix_mtx.tocsc()[:, cell_indices].tocsr()
 
 # select 'num_genes_to_keep' highly expressed genes
-genes_to_keep_indices = sorted(range(num_raw_genes),
-    key=lambda x: total_gene_expression[x], reverse=True)[:num_genes_to_keep]
+total_gene_expression = np.array(adata.X.sum(axis=0)).squeeze()
+genes_to_keep_indices = np.argsort(total_gene_expression)[::-1][:num_genes_to_keep]
+
+# slice dataset on genes
+adata = adata[:, genes_to_keep_indices].copy()
+
+# find cells and empties
+umi_per_barcode = np.array(adata.X.sum(axis=1)).squeeze()
+umi_sorted_barcode_indices = np.argsort(umi_per_barcode)[::-1]
+cell_indices = umi_sorted_barcode_indices[:expected_cells]
+empty_indices = umi_sorted_barcode_indices[expected_cells:]
 
 # putative list of barcodes to keep
 cell_barcodes_to_keep_indices = np.asarray(cell_indices)[
@@ -70,34 +61,21 @@ empty_barcodes_to_keep_indices = np.asarray(empty_indices)[
     rng.permutation(len(empty_indices))[:num_empty_barcodes_to_keep]].tolist()
 barcodes_to_keep_indices = cell_barcodes_to_keep_indices + empty_barcodes_to_keep_indices
 
-# remove barcodes with zero expression on kept genes
-trimmed_counts_matrix = matrix_mtx.tocsr()[genes_to_keep_indices, :].tocsc()[:, barcodes_to_keep_indices]
-umi_per_putatively_kept_barcodes = np.asarray(np.sum(trimmed_counts_matrix, 0)).flatten()
-barcodes_to_keep_indices = np.asarray(barcodes_to_keep_indices)[umi_per_putatively_kept_barcodes > 0]
-barcodes_to_keep_indices = barcodes_to_keep_indices.tolist()
-
-# slice the raw dataset
-tiny_matrix_mtx = matrix_mtx.tocsr()[genes_to_keep_indices, :].tocsc()[:, barcodes_to_keep_indices].tocoo()
-tiny_genes_df = genes_df.loc[genes_to_keep_indices]
-tiny_barcodes_df = barcodes_df.loc[barcodes_to_keep_indices]
+# slice dataset on barcodes
+adata = adata[barcodes_to_keep_indices].copy()
 
 # compensate for lost counts (due to the reduced number of genes)
-tiny_matrix_mtx = tiny_matrix_mtx * int((num_raw_genes / num_genes_to_keep) ** 0.25)
+adata.X = adata.X * int((num_raw_genes / num_genes_to_keep) ** 0.25)
 
 print(f"Number of genes in the trimmed dataset: {len(genes_to_keep_indices)}")
 print(f"Number of barcodes in the trimmed dataset: {len(barcodes_to_keep_indices)}")
 print(f"Expected number of cells in the trimmed dataset: {num_cell_barcodes_to_keep}")
 
-# save
-print("Saving the trimmed dataset...")
-output_path = os.path.join(os.curdir, 'tiny_raw_gene_bc_matrices', 'GRCh38')
-os.makedirs(output_path, exist_ok=True)
-mmwrite(os.path.join(output_path, "matrix.mtx"), tiny_matrix_mtx)
-tiny_genes_df.to_csv(os.path.join(output_path, "genes.tsv"), sep='\t', header=None, index=False)
-tiny_barcodes_df.to_csv(os.path.join(output_path, "barcodes.tsv"), sep='\t', header=None, index=False)
+print(adata)
 
-print("Cleaning up...")
-shutil.rmtree(extracted_dataset_local_path, ignore_errors=True)
-os.remove(dataset_local_filename)
+# save
+output_file = 'tiny_raw_gene_bc_matrices_h5.h5ad'
+print(f"Saving the trimmed dataset as {output_file} ...")
+adata.write(output_file)
 
 print("Done!")
